third_party/blink/web_tests/svg/dynamic-updates/SVGUseElement-dom-href2-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'href' attribute of the SVGUseElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS useElement.getAttributeNS('http://www.w3.org/1999/xlink', 'href') is "#MyRect"
PASS successfullyParsed is true

TEST COMPLETE
